[PC-515] Chore: 리스폰스 파일 분리
